backend/requirements.txt
# Runtime dependencies (direct)
fastapi==0.121.3
pydantic==2.12.4
pydantic-settings==2.11.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.2
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.20
redis==7.1.0
slowapi==0.1.9
boto3==1.41.5
resend==2.4.0
uvicorn[standard]==0.38.0
httpx==0.28.1

# Development / verification dependencies
pytest==9.0.1
pytest-asyncio==1.3.0
ruff==0.14.6
mypy==1.18.2
